@@ -12,14 +12,14 @@
 		"deploy": "touch dist/.nojekyll && gh-pages -d dist -t true"
 	},
 	"devDependencies": {
-		"@sveltejs/vite-plugin-svelte": "^3.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@tsconfig/svelte": "^4.0.1",
 		"gh-pages": "^5.0.0",
-		"svelte": "^4.0",
-		"svelte-check": "^3.6.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tslib": "^2.6.0",
-		"typescript": "^5.0",
-		"vite": "^5.0",
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4",
 		"vitest": "^1"
 	},
 	"dependencies": {
