@@ -1,35 +1,43 @@
 <script lang="ts">
-    import { createEventDispatcher } from 'svelte';
-    import { NUM_COLORS } from '../game/game.js';
+    import { NUM_COLORS } from "../game/game.js";
 
-    const dispatch = createEventDispatcher<{'color-pick': number}>();
+    interface Props {
+        counts?: { [key: number]: number };
+        onColorPick: (color: number) => void;
+    }
 
-    export let counts: {[key: number]: number} = {}
-    
-    const colors = Array(NUM_COLORS).fill(null).map((_, i) => i + 1)
-    let selected: number = null
+    let { counts = {}, onColorPick }: Props = $props();
+
+    const colors = Array(NUM_COLORS)
+        .fill(null)
+        .map((_, i) => i + 1);
+    let selected: number = $state(null);
 
     function selectColor(color: number) {
-        selected = color
-        dispatch('color-pick', color)
+        selected = color;
+        onColorPick(color);
     }
 </script>
 
-
 <div class="palette">
     {#each colors as color}
-        <button on:click={() => selectColor(color)}
+        <button
+            onclick={() => selectColor(color)}
             class="cell"
-            class:curr={color === selected} 
+            class:curr={color === selected}
             class:extra={counts[color] > 4}
-            style:--color="var(--clr-wtr{color >= 0 ? color : '-none'})">
-            {counts[color] || ''}
+            style:--color="var(--clr-wtr{color >= 0 ? color : '-none'})"
+        >
+            {counts[color] || ""}
         </button>
     {/each}
-    <button on:click={() => selectColor(null)}
+    <button
+        onclick={() => selectColor(null)}
         class="cell empty"
-        class:curr={null === selected} 
-        style:--color="var(--clr-wtr-none)">
+        class:curr={null === selected}
+        style:--color="var(--clr-wtr-none)"
+        aria-label="No Color"
+    >
     </button>
 </div>
 
@@ -41,35 +49,36 @@
     .cell {
         width: 2.5rem;
         height: 1.8rem;
-        border:1px solid white;
+        border: 1px solid white;
         display: inline-block;
         background-color: var(--color);
         position: relative;
     }
     .curr {
-        outline: black solid .125rem;
-        outline-offset: -.25rem;
+        outline: black solid 0.125rem;
+        outline-offset: -0.25rem;
     }
     .curr.empty {
-        outline: white solid .125rem;
-        outline-offset: -.25rem;
+        outline: white solid 0.125rem;
+        outline-offset: -0.25rem;
     }
-    .empty:before,.empty:after {
-        content: '';
+    .empty:before,
+    .empty:after {
+        content: "";
         width: 75%;
-        height: .125rem;
+        height: 0.125rem;
         background-color: white;
         position: absolute;
         top: 50%;
         left: 50%;
-        transform: translateX(-50%) rotate(-36deg) ;
+        transform: translateX(-50%) rotate(-36deg);
     }
-    .empty:before{
-        transform:  translateX(-50%) rotate(36deg);
-        
+    .empty:before {
+        transform: translateX(-50%) rotate(36deg);
     }
     .extra {
-        outline: red solid .25rem;
-        outline-offset: -.25rem;
+        outline: red solid 0.25rem;
+        outline-offset: -0.25rem;
     }
 </style>
+
